@@ -1,63 +1,51 @@
 [project]
 name="new-modeling-toolkit"
 description="E3's New Modeling Toolkit"
-version="0.17.4"
+version="3.5.2"
 authors=[
     { name = "Energy and Environmental Economics, Inc." }
 ]
-requires-python = ">=3.9,<3.11"
-dependencies = [
-    "fastparquet==2023.10.1",
-    "highspy==1.5.3",
-    "ipywidgets==8.0.4",
-    "joblib==1.2.0",
-    "jupyterlab==3.5.3",
-    "jupytext==1.15.1",
-    "kaleido==0.2.1",
-    "kmedoids==0.4.3",
-    "line-profiler==4.0.3",
-    "loguru==0.6.0",
-    "matplotlib==3.6.3",
-    "memory-profiler==0.61.0",
-    "numpy==1.24.1",
-    "pandas==1.5.3",
-    "pint==0.20.1",
-    "plotly==5.13.0",
-    "pyarrow==14.0.1",
-    "pydantic==1.10.8",
-    "pyomo==6.4.4",
-    "rich==13.7.0",
-    "s3fs==2023.9.0",
-    "scikit-learn==1.2.1",
-    "scipy==1.10.0",
-    "tqdm==4.64.1",
-    "typer==0.7.0",
-    "ujson==5.8.0",
-    "universal-pathlib==0.1.3",
-]
+requires-python = ">=3.10"
+
+dynamic = [ "dependencies" ]
 
 [project.optional-dependencies]
+jupyter = [
+    "bokeh >= 3.1.0",
+    "ipywidgets==8.1.3",
+    "openpyxl==3.1.5",
+    "jupyterlab==3.6.7",
+    "jupytext==1.16.2",
+]
 ui = [
     "ipyfilechooser==0.6.0",
-    "python-pptx==0.6.21",
-    "streamlit==1.28.2",
-    "xlwings==0.29.0",
+    "python-pptx==0.6.23",
+    "streamlit==1.35.0",
 ]
 dev = [
-    "autodoc_pydantic==1.9.0",
-    "commitizen==3.13.0",
-    "darker==1.7.2",
-    "furo==2023.9.10",
-    "graphviz==0.20.1",
+    "autodoc_pydantic==2.2.0",
+    "commitizen==3.25.0",
+    "darker==1.7.3",
+    "graphviz==0.20.3",
+    "line-profiler==4.1.3",
+    "memory-profiler==0.61.0",
     "myst-parser==2.0.0",
-    "pre-commit==3.6.0",
+    "pre-commit==3.7.1",
+    "pylint==3.3.5",
     "pytest==7.4.4",
     "pytest-sugar==0.9.7",
-    "pytest-xdist[psutil]==3.5.0",
+    "pytest-xdist[psutil]==3.6.1",
     "sphinx==7.2.6",
-    "sphinx-copybutton==0.5.2",
+    "sphinxawesome-theme==5.1.4",
+    "sphinxcontrib-mermaid==0.9.2",
+    "sphinx-autobuild==2024.4.16",
     "sphinx-design==0.5.0",
+    "sphinx-last-updated-by-git==0.3.7",
 ]
+
+[project.scripts]
+nmt = "new_modeling_toolkit.cli:main"
+kit-ui = "new_modeling_toolkit.cli.new:main"
 
 [tool.setuptools.packages.find]
 where = ["."]
@@ -65,6 +53,7 @@ include = ["new_modeling_toolkit*"]
 
 [tool.setuptools.package-data]
 "new_modeling_toolkit.common" = ["units.csv"]
+"new_modeling_toolkit.ui" = ["runTerminalCommand-0.24.0.applescript"]
 
 [tool.black]
 line_length = 120
@@ -75,4 +64,7 @@ tag_format = "$version"
 version_type = "pep440"
 version_provider = "pep621"
 update_changelog_on_bump = true
-major_version_zero = true
+major_version_zero = false
+
+[tool.setuptools.dynamic]
+dependencies = { file = [ "requirements.txt" ] }
