@@ -1,19 +1,29 @@
+from __future__ import annotations
+
 import enum
 import os
 import pathlib
-import typing
+import textwrap
+from typing import ClassVar
 from typing import List
+from typing import Literal
 from typing import Optional
 
+import numpy as np
 import pandas as pd
 import pydantic
 from loguru import logger
+from pydantic import Field
 from tqdm import tqdm
+from typing_extensions import Annotated
 
-from new_modeling_toolkit import get_units
 from new_modeling_toolkit.core import component
-from new_modeling_toolkit.core.custom_model import convert_str_float_to_int
+from new_modeling_toolkit.core.custom_model import get_units
+from new_modeling_toolkit.core.custom_model import Metadata
+from new_modeling_toolkit.core.custom_model import units
 from new_modeling_toolkit.core.temporal import timeseries as ts
+from new_modeling_toolkit.core.utils.pandas_utils import compare_dataframes
+from new_modeling_toolkit.core.utils.xlwings import ExcelApiCalls
 
 
 @enum.unique
@@ -42,33 +52,57 @@ class Linkage(component.Component):
     # CLASS ATTRIBUTES #
     ####################
 
-    _RELATIONSHIP_TYPE: LinkageRelationshipType
+    _RELATIONSHIP_TYPE: ClassVar[LinkageRelationshipType]
 
-    _instances: typing.ClassVar = {}
-    _component_type_from: typing.ClassVar = None
-    _component_type_to: typing.ClassVar = None
-    _class_descriptor: typing.ClassVar = ""  # This is the name for printing the info message
-    _instance_from: component.Component
-    _instance_to: component.Component
+    component_type_from_: str
+    component_type_to_: str
+    instance_from: component.Component | None = (
+        None  # Not ideal, since a linkage should only exist if there are linked instances
+    )
+    instance_to: component.Component | None = None
     # Filename attributes
-    _attribute_file: typing.ClassVar = None
+    SAVE_PATH: ClassVar[str | None] = None
 
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
 
-        # Append new instance to the `_instances` class attribute
-        if self.__class__.__name__ not in Linkage._instances.keys():
-            Linkage._instances.update({self.__class__.__name__: [self]})
+    def __eq__(self, other):
+        if not isinstance(other, Linkage):
+            equals = False
         else:
-            Linkage._instances[self.__class__.__name__].append(self)
+            equals = (
+                type(self) == type(other)
+                and self.name == other.name
+                and self.component_type_to_ == other.component_type_to_
+                and self.component_type_from_ == other.component_type_from_
+                and id(self.instance_from) == id(other.instance_from)
+                and id(self.instance_to) == id(other.instance_from)
+            )
+
+        return equals
 
     def dict(self, **kwargs):
-        """Need to exclude `_instance_from`, `_instance_to` attributes to avoid recursion error when saving to JSON."""
-        attrs_to_exclude = {"attr_path", "_instance_from", "_instance_to", "_component_type_from", "_component_type_to"}
+        """Need to exclude `instance_from`, `instance_to` attributes to avoid recursion error when saving to JSON."""
+        attrs_to_exclude = {"attr_path", "instance_from", "instance_to", "component_type_from_", "component_type_to_"}
 
-        dct = super(Linkage, self).dict(exclude=attrs_to_exclude, exclude_defaults=True, exclude_none=True)
+        dct = super(Linkage, self).model_dump(exclude=attrs_to_exclude, exclude_defaults=True, exclude_none=True)
 
         return dct
+
+    @pydantic.field_validator("name")
+    def _validate_name_tuple(cls, name):
+        if not all([type(name) is tuple, len(name) == 2, all(type(s) is str for s in name)]):
+            raise AssertionError("Linkage names must be a tuple of 2 strings")
+
+        return name
+
+    @property
+    def name_from(self):
+        return self.name[0]
+
+    @property
+    def name_to(self):
+        return self.name[1]
 
     @classmethod
     def _filter_scenarios(cls, linkages_df: pd.DataFrame, scenarios: List[str], filepath: pathlib.Path) -> pd.DataFrame:
@@ -103,18 +137,85 @@ class Linkage(component.Component):
         return filtered_linkages_df
 
     @classmethod
+    def to_excel(cls, *, anchor_range, excel_api: ExcelApiCalls, **kwargs):
+        """Small wrapper around Component.to_excel() because for now linkages need a formula to construct their name."""
+        rows_to_add, columns_to_add, table_name = super().to_excel(
+            anchor_range=anchor_range, excel_api=excel_api, **kwargs
+        )
+
+        # Add formula for linkage name
+        name_column = anchor_range.sheet.range(f"{table_name}[Name]")
+        excel_api.set_cell_style(name_column, excel_api.styles(name_column.sheet.book, "Calculation"))
+        name_column.value = f'=TEXTJOIN(", ", TRUE, {table_name}[@[Instance From]:[Instance To]])'
+
+        return rows_to_add, columns_to_add, table_name
+
+    @classmethod
+    def dfs_to_csv(
+        cls,
+        *,
+        instances: pd.DataFrame,
+        wb: "Book",
+        dir_str: "DirStructure",
+        compare_files: bool = True,
+        dry_run: bool = False,
+    ) -> None:
+        """Save DataFrame into (separate) component CSV files."""
+        save_path = dir_str.data_interim_dir / "linkages"
+        save_path.mkdir(parents=True, exist_ok=True)
+
+        # Split the name into `component_from`, `component_to` (assumes that the name column is a formula)
+        instances = instances.loc[~instances["attribute"].isin(["instance_from", "instance_to", "to", "from"]), :]
+        if instances.empty:
+            return None
+
+        id = instances["name"].str.split(", ", expand=True)
+        id.columns = ["component_from", "component_to"]
+        instances = pd.concat([instances.drop("name", axis=1), id], axis=1, ignore_index=True)
+
+        # Reorder columns
+        instances.columns = ["attribute", "timestamp", "value", "scenario", "component_from", "component_to"]
+        instances = instances[["component_from", "component_to", "attribute", "timestamp", "value", "scenario"]]
+
+        if (
+            (save_path / f"{cls.SAVE_PATH}").exists()
+            and os.stat(save_path / f"{cls.SAVE_PATH}").st_size > 0
+            and compare_files
+        ):
+            try:
+                previous_df = pd.read_csv(
+                    save_path / f"{cls.SAVE_PATH}", parse_dates=["timestamp"], infer_datetime_format=True
+                )
+                comparison = compare_dataframes(
+                    previous=previous_df,
+                    new=instances,
+                    indices=["component_from", "component_to", "attribute", "timestamp", "scenario"],
+                    column_to_compare="value",
+                )
+
+                if not comparison.empty:
+                    comparison = textwrap.indent(comparison.to_string(), " > ")
+                    logger.debug(f'│ ├─ Differences in {save_path / f"{cls.SAVE_PATH}"}:\n{comparison}')
+            except:
+                logger.error(f'Could not compare CSV files: {save_path / f"{cls.SAVE_PATH}"}.csv')
+
+        if not dry_run:
+            instances.to_csv(save_path / f"{cls.SAVE_PATH}", index=False)
+            wb.app.status_bar = f"Writing {cls.__name__}"
+
+    @classmethod
     def from_dir(
         cls,
         dir_path: os.PathLike,
         linkages_df: pd.DataFrame,
-        components_dict: dict,
+        components_dict: dict[str, component.Component],
         linkages_csv_path: pathlib.Path,
         scenarios: list = [],
     ):
         """Create Linkage instances based on prescribed CSV files.
 
         This method relies on one class attribute:
-            _attribute_file: Optional filename for static & dynamic attributes of the Linkage instance
+            SAVE_PATH: Optional filename for static & dynamic attributes of the Linkage instance
 
         Args:
             dir_path: Path to CSV file
@@ -130,68 +231,65 @@ class Linkage(component.Component):
         )
 
         # TODO (5/13): from_csv shares similarities with Component.from_csv...maybe want to generalize both later
-        # Read in _attribute_file data as needed
+        # Read in SAVE_PATH data as needed
         linkage_attributes = None
-        if cls._attribute_file is not None:  # Read in CSV attributes file if it exists
-            input_df = pd.read_csv(pathlib.Path(dir_path) / cls._attribute_file)
+        if (
+            cls.SAVE_PATH is not None and (pathlib.Path(dir_path) / cls.SAVE_PATH).exists()
+        ):  # Read in CSV attributes file if it exists
+            input_df = pd.read_csv(pathlib.Path(dir_path) / cls.SAVE_PATH)
+            input_df["timestamp"] = input_df["timestamp"].fillna(value="None")
             linkage_attributes = cls._filter_scenarios(
-                linkages_df=input_df, scenarios=scenarios, filepath=pathlib.Path(dir_path) / cls._attribute_file
+                linkages_df=input_df, scenarios=scenarios, filepath=pathlib.Path(dir_path) / cls.SAVE_PATH
             )
-            linkage_attributes = linkage_attributes.groupby(["component_from", "component_to"])
+            if len(linkage_attributes) == 0:
+                linkage_attributes = None
+            else:
+                linkage_attributes = linkage_attributes.groupby(["component_from", "component_to"])
 
         # Construct linkage instances
+        constructed_linkages = []
         unmatched_linkages = []
         for name_from, name_to in tqdm(
             linkage_pairs,
-            desc=f"Loading {cls.__name__}".rjust(32),
+            desc=f"Loading {cls.__name__}".rjust(50),
             bar_format="{l_bar}{bar:30}{r_bar}{bar:-10b}",
         ):
-            # For `AllToPolicy` class the component_type from is an input, for all the other classes it is a class
-            # level static attrs
-            if cls == AllToPolicy:
-                component_type_from = find_component_type(components_dict, name_from)
-            else:
-                component_type_from = cls._component_type_from
-            component_type_to = cls._component_type_to
 
             # obtain the linked inst from and to from the components dictionary
-            instance_from = (
-                components_dict[component_type_from].get(name_from, None) if component_type_from is not None else None
-            )
-            instance_to = components_dict[component_type_to].get(name_to, None)
+            instance_from = components_dict.get(name_from, None)
+            instance_to = components_dict.get(name_to, None)
 
             if instance_from is None or instance_to is None:
-                unmatched_linkages += [f"{cls.__name__}({name_from}, {name_to})"]
+                missing_linkage = f"{cls.__name__}({name_from}, {name_to}): missing {name_from if instance_from is None else ' '} {name_to if instance_to is None else ' '}"
+                unmatched_linkages += [missing_linkage]
+
             else:
                 if linkage_attributes is None or (name_from, name_to) not in linkage_attributes.groups:
                     linkage_instance = cls(
-                        name=(name_from, name_to), _instance_from=instance_from, _instance_to=instance_to
+                        name=(name_from, name_to), instance_from=instance_from, instance_to=instance_to
                     )
                 else:
-                    _, linkage_instance = cls._parse_vintages(
-                        filename=pathlib.Path(dir_path) / cls._attribute_file,
+                    linkage_instance = cls.from_dataframe(
                         input_df=linkage_attributes.get_group((name_from, name_to)).drop(
                             columns=["component_from", "component_to"]
                         ),
-                        separate_vintages=False,
+                        attr_path=pathlib.Path(dir_path) / cls.SAVE_PATH,
                         scenarios=scenarios,
-                        data={"_instance_from": instance_from, "_instance_to": instance_to},
+                        data={"instance_from": instance_from, "instance_to": instance_to},
                         name=(name_from, name_to),
-                    ).popitem()
+                    )
 
-                # This ensures _component_type_from is correct.
-                if cls.__name__ == "AllToPolicy" and component_type_from == "assets":
-                    linkage_instance._component_type_from = "assets_"
-                else:
-                    linkage_instance._component_type_from = component_type_from
+                constructed_linkages.append(linkage_instance)
 
         if len(unmatched_linkages) > 0:
-            logger.warning(
-                f"The following linkages were not loaded (could not find corresponding components in system): \n{unmatched_linkages}"
+            formatted_list = "\n\t".join(unmatched_linkages)
+            logger.debug(
+                f"The following linkages were not loaded (could not find corresponding components in system): \n\t{formatted_list}"
             )
 
-    @classmethod
-    def announce_linkage_to_instances(cls):
+        return constructed_linkages
+
+    def announce_linkage_to_instances(self):
         """Iterate through all linkages in Linkages._instances to append instances as attributes to mapped instances.
 
         In other words, if we have:
@@ -206,66 +304,28 @@ class Linkage(component.Component):
         So that it's easy to find the necessary attributes from linked instances
         (e.g., fuel burn associated with a fuel being burned in a certain resource)
         """
-        for linkage_type, linkage_instances in cls._instances.items():
-            for linkage in tqdm(
-                linkage_instances,
-                desc=f"Loading {cls.__name__}".rjust(32),
-                bar_format="{l_bar}{bar:30}{r_bar}{bar:-10b}",
-            ):
-                logger.debug(
-                    f"Announcing linkage between '{linkage._instance_from.name}', '{linkage._instance_to.name}'"
-                )
-                # Unpack the tuple
-                name_from, name_to = linkage.name
-                # instance to look at, attribute to look at & append opposing component instance, name of opposing component instance
-                linkage_tuple = [
-                    (linkage._instance_to, linkage._component_type_from, name_from),
-                    (linkage._instance_from, linkage._component_type_to, name_to),
-                ]
-                # TODO (2021-11-16): Related to #380, can simplify this (if default linkage attribute is {} instead of None)
-                for instance, attr, name in linkage_tuple:
-                    if attr is not None:
-                        if getattr(instance, attr) is None:
-                            # Create a new dict
-                            instance.__dict__[attr] = {name: linkage}
-                        else:
-                            # Update the dict with additional values
-                            instance.__dict__[attr].update({name: linkage})
+        # Unpack the tuple
+        # instance to look at, attribute to look at & append opposing component instance, name of opposing component instance
+        linkage_tuple = [
+            (self.instance_to, self.component_type_from_, self.name_from),
+            (self.instance_from, self.component_type_to_, self.name_to),
+        ]
+        # TODO (2021-11-16): Related to #380, can simplify this (if default linkage attribute is {} instead of None)
+        for instance, attr, name in linkage_tuple:
+            if attr is not None and attr != "":
+                if getattr(instance, attr) is None:
+                    # Create a new dict
+                    instance.__dict__[attr] = {name: self}
+                else:
+                    # Update the dict with additional values
+                    if attr in instance.__dict__.keys():
+                        instance.__dict__[attr].update({name: self})
+                    else:
+                        instance.__dict__[attr] = {}
+                        instance.__dict__[attr].update({name: self})
 
-    @classmethod
-    def save_instance_attributes_csvs(cls, wb, data: pd.DataFrame, save_path: pathlib.Path, overwrite: bool = True):
-        """Save DataFrame in ``attributes.csv`` format, splitting DataFrame into separate CSVs for each instance.
-
-        This method is a variation on ``Component.save_instance_attributes_csvs`` because:
-            1. ``Linkage`` instances are saved in the same CSV file; whereas ``Components`` have separate ``attributes.csv`` files for each instance.
-            2. ``Linkage`` instances are indexed by the names of the components that are linked together.
-
-        Args:
-            wb: An ``xlwings`` workbook (only used to method prints a progress message)
-            df: Combined DataFrame in "long" ``attributes.csv`` format (see ``cls.get_data_from_xlwings``).
-            save_folder: Path to folder that will hold CSV files.
-            overwrite: Whether this method should overwrite an existing attributes.csv file. Otherwise, will append unique values to existing file.
-                # TODO 2022-05-05: Add this overwrite feature
-
-        Returns:
-
-        """
-        data = (
-            data.reset_index()
-            .dropna(subset="value")
-            .rename(
-                columns={
-                    "level_0": "component_from",
-                    "level_1": "component_to",
-                    "level_2": "scenario",
-                }
-            )
-        )[["component_from", "component_to", "timestamp", "attribute", "value", "scenario"]]
-
-        wb.app.status_bar = f"Writing {cls.__name__}"
-
-        save_path.parent.mkdir(parents=True, exist_ok=True)
-        data.to_csv(save_path, index=False)
+    def copy(self):
+        return self.model_copy()
 
 
 class HybridStorageResourceToHybridVariableResource(Linkage):
@@ -274,17 +334,39 @@ class HybridStorageResourceToHybridVariableResource(Linkage):
     ####################
     _RELATIONSHIP_TYPE = LinkageRelationshipType.ONE_TO_ONE
     _class_descriptor = "hybrid_storage_resources → hybrid_variable_resources"
-    _component_type_from = "hybrid_storage_resources"
-    _component_type_to = "hybrid_variable_resources"
-    _attribute_file = "hybrid_resources.csv"
+    component_type_from_: str = "hybrid_storage_resources"
+    component_type_to_: str = "hybrid_variable_resources"
+    SAVE_PATH = "hybrid_resources.csv"
 
-    interconnection_limit_mw: ts.NumericTimeseries = pydantic.Field(
-        default_freq="YS",
-        up_method="ffill",
-        down_method="average",
+    interconnection_limit_mw: ts.NumericTimeseries | None = pydantic.Field(
+        default=None, default_freq="YS", up_method="ffill", down_method="mean"
     )
 
-    grid_charging_allowed: bool = False
+    grid_charging_allowed: Optional[bool] = Field(
+        False,
+        description="If True, hybrid_storage's power_input must not exceed hybrid_variable's power_output.",
+    )
+
+    pairing_ratio: Optional[float] = Field(
+        None,
+        description="If specified, ratio of operational capacity of hybrid_variable to that of hybrid_storage resources. Usually set to 1.",
+        ge=0,
+    )
+
+    paired_charging_constraint_active_in_year: ts.NumericTimeseries = pydantic.Field(
+        description="If 1, the charging constraint is active in the year. If 0, the charging constraint is not active in the year.",
+        default_factory=ts.Timeseries.one,
+        default_freq="YS",
+        up_method="ffill",
+        down_method="mean",
+    )
+
+    def revalidate(self):
+        if any(~self.paired_charging_constraint_active_in_year.data.isin([0, 1])):
+            return ValueError(
+                f"Values in `paired_charging_constraint_active_in_year` attribute for `{self.__class__.__name__}` "
+                f"instance `{self.name}` can only be either 1 (constraint active) or 0 (constraint inactive)."
+            )
 
 
 class CandidateFuelToResource(Linkage):
@@ -298,8 +380,8 @@ class CandidateFuelToResource(Linkage):
     ####################
     _RELATIONSHIP_TYPE = LinkageRelationshipType.MANY_TO_MANY
     _class_descriptor = "candidate fuels → resources"
-    _component_type_from = "candidate_fuels"
-    _component_type_to = "resources"
+    component_type_from_: str = "candidate_fuels"
+    component_type_to_: str = "resources"
 
     ###################
     # INSTANCE FIELDS #
@@ -320,15 +402,16 @@ class ResourceToReserve(Linkage):
     ####################
     _RELATIONSHIP_TYPE = LinkageRelationshipType.MANY_TO_MANY
     _class_descriptor = "resources → operating reserves"
-    _component_type_from = "resources"
-    _component_type_to = "reserves"
-    _attribute_file = "resources_to_reserves.csv"
+    component_type_from_: str = "resources"
+    component_type_to_: str = "reserves"
+    SAVE_PATH = "resources_to_reserves.csv"
 
     ###################
     # INSTANCE FIELDS #
     ###################
-    exclusive: bool = True
-    dependent_on: str = pydantic.Field(
+    # TODO: Are `exclusive` and `dependent_on` being used?
+    exclusive: Annotated[bool, Metadata(default_exclude=True)] = True
+    dependent_on: Annotated[str, Metadata(default_exclude=True)] = pydantic.Field(
         "setpoint",
         description="Operating reserves are usually related to a resource's setpoint (i.e., online head/footroom). "
         "Certain resources (e.g., inertia) are only related to online/committed capacity."
@@ -336,38 +419,38 @@ class ResourceToReserve(Linkage):
     )
     # TODO (2022-02-21): Non-spinning reserves cannot be represented with these options, since they can also include offline capacity.
 
-    max_fraction_of_capacity: pydantic.condecimal(ge=0, le=1) = pydantic.Field(
-        1,
+    max_fraction_of_capacity: Annotated[float, Field(ge=0, le=1)] = pydantic.Field(
+        0,
         description="Max % of a resource's online capacity (e.g., committed capacity for unit commitment resources) "
         "that can be used to provide operating reserve.",
     )
-    incremental_requirement_annual_scalar: typing.Optional[ts.NumericTimeseries] = pydantic.Field(
-        None,
-        default_freq="YS",
-        up_method="interpolate",
-        down_method="first",
-        units=get_units("requirement"),
-    )
     scalar_type: IncrementalReserveType = IncrementalReserveType.OPERATIONAL_CAPACITY
 
-    incremental_requirement_hourly_scalar: typing.Optional[ts.NumericTimeseries] = pydantic.Field(
-        None,
+    incremental_requirement_hourly_scalar: ts.NumericTimeseries = pydantic.Field(
         default_freq="H",
         up_method="interpolate",
         down_method="first",
-        units=get_units("requirement"),
         alias="incremental_requirement",
+        default_factory=ts.NumericTimeseries.zero,
     )
     incremental_requirement_hourly_scalar__type: ts.TimeseriesType = ts.TimeseriesType.MODELED_YEAR
 
-    @pydantic.validator("dependent_on")
-    def validate_reserve_dependence(cls, dependent_on, values):
-        dependent_list = ["setpoint", "commitment"]
-        if dependent_on not in dependent_list:
-            raise ValueError(
-                f"For {cls.__name__} linkage \"{values['name']}\", attribute (`dependent_on`) must be in {dependent_list}"
-            )
-        return dependent_on
+    # TODO: incremental requirement annual scalar is currently not used
+    incremental_requirement_annual_scalar: ts.NumericTimeseries = pydantic.Field(
+        default_freq="YS",
+        up_method="interpolate",
+        down_method="first",
+        default_factory=ts.NumericTimeseries.zero,
+    )
+
+    # @pydantic.validator("dependent_on")
+    # def validate_reserve_dependence(cls, dependent_on, values):
+    #     dependent_list = ["setpoint", "commitment"]
+    #     if dependent_on not in dependent_list:
+    #         raise ValueError(
+    #             f"For {cls.__name__} linkage \"{values['name']}\", attribute (`dependent_on`) must be in {dependent_list}"
+    #         )
+    #     return dependent_on
 
 
 class LoadToZone(Linkage):
@@ -376,8 +459,8 @@ class LoadToZone(Linkage):
     ####################
     _RELATIONSHIP_TYPE = LinkageRelationshipType.MANY_TO_ONE
     _class_descriptor = "loads → zones"
-    _component_type_from = "loads"
-    _component_type_to = "zones"
+    component_type_from_: str = "loads"
+    component_type_to_: str = "zones"
 
     ###################
     # INSTANCE FIELDS #
@@ -389,17 +472,17 @@ class LoadToReserve(Linkage):
     ####################
     # CLASS ATTRIBUTES #
     ####################
-    _RELATIONSHIP_TYPE = LinkageRelationshipType.ONE_TO_MANY
+    _RELATIONSHIP_TYPE = LinkageRelationshipType.MANY_TO_MANY
     _class_descriptor = "loads → reserves"
-    _component_type_from = "loads"
-    _component_type_to = "reserves"
-    _attribute_file = "loads_to_reserves.csv"
+    component_type_from_: str = "loads"
+    component_type_to_: str = "reserves"
+    SAVE_PATH = "loads_to_reserves.csv"
 
     ###################
     # INSTANCE FIELDS #
     ###################
-    incremental_requirement_hourly_scalar: Optional[ts.FractionalTimeseries] = pydantic.Field(
-        default=None,
+    incremental_requirement_hourly_scalar: ts.FractionalTimeseries = pydantic.Field(
+        default_factory=ts.NumericTimeseries.zero,
         default_freq="H",
         up_method="ffill",
         down_method="average",
@@ -408,22 +491,21 @@ class LoadToReserve(Linkage):
 
 
 class ReserveToZone(Linkage):
-    """TODO: 2023-08-22: I want to deprecate this class in favor of `LoadToReserve`, but for now keeping for backward compatibility."""
 
     ####################
     # CLASS ATTRIBUTES #
     ####################
     _RELATIONSHIP_TYPE = LinkageRelationshipType.MANY_TO_ONE
     _class_descriptor = "reserves → zones"
-    _component_type_from = "reserves"
-    _component_type_to = "zones"
-    _attribute_file = "reserves_to_zones.csv"
+    component_type_from_: str = "reserves"
+    component_type_to_: str = "zones"
+    SAVE_PATH = "reserves_to_zones.csv"
 
     ###################
     # INSTANCE FIELDS #
     ###################
-    incremental_requirement_hourly_scalar: Optional[ts.FractionalTimeseries] = pydantic.Field(
-        default=None,
+    incremental_requirement_hourly_scalar: ts.FractionalTimeseries = pydantic.Field(
+        default_factory=ts.NumericTimeseries.zero,
         default_freq="H",
         up_method="ffill",
         down_method="average",
@@ -438,8 +520,8 @@ class ResourceToZone(Linkage):
     ####################
     _RELATIONSHIP_TYPE = LinkageRelationshipType.MANY_TO_ONE
     _class_descriptor = "resources → zones"
-    _component_type_from = "resources"
-    _component_type_to = "zones"
+    component_type_from_: str = "resources"
+    component_type_to_: str = "zones"
 
     ###################
     # INSTANCE FIELDS #
@@ -449,14 +531,14 @@ class ResourceToZone(Linkage):
 
 class AssetToAssetGroup(Linkage):
     _RELATIONSHIP_TYPE = LinkageRelationshipType.MANY_TO_MANY
-    _component_type_from = "assets"
-    _component_type_to = "asset_groups"
+    component_type_from_: str = "assets"
+    component_type_to_: str = "asset_groups"
 
 
 class TrancheToAsset(Linkage):
     _RELATIONSHIP_TYPE = LinkageRelationshipType.MANY_TO_ONE
-    _component_type_from = "tranches"
-    _component_type_to = "assets"
+    component_type_from_: str = "tranches"
+    component_type_to_: str = "assets"
 
 
 class ResourceToResourceGroup(Linkage):
@@ -465,8 +547,8 @@ class ResourceToResourceGroup(Linkage):
     ####################
     _RELATIONSHIP_TYPE = LinkageRelationshipType.MANY_TO_ONE
     _class_descriptor = "resources → resource groups"
-    _component_type_from = "resources"
-    _component_type_to = "resource_groups"
+    component_type_from_: str = "resources"
+    component_type_to_: str = "resource_groups"
 
     ###################
     # INSTANCE FIELDS #
@@ -480,8 +562,8 @@ class ResourceToOutageDistribution(Linkage):
     ####################
     _RELATIONSHIP_TYPE = LinkageRelationshipType.MANY_TO_ONE
     _class_descriptor = "resources → outage distributions"
-    _component_type_from = "resources"
-    _component_type_to = "outage_distributions"
+    component_type_from_: str = "resources"
+    component_type_to_: str = "outage_distributions"
 
     ###################
     # INSTANCE FIELDS #
@@ -495,8 +577,8 @@ class ResourceToFuelZone(Linkage):
     ####################
     _RELATIONSHIP_TYPE = LinkageRelationshipType.MANY_TO_ONE
     _class_descriptor = "resources → zones"
-    _component_type_from = "resources"
-    _component_type_to = "fuel_zones"
+    component_type_from_: str = "resources"
+    component_type_to_: str = "fuel_zones"
 
     ###################
     # INSTANCE FIELDS #
@@ -510,8 +592,8 @@ class AssetToZone(Linkage):
     ####################
     _RELATIONSHIP_TYPE = LinkageRelationshipType.MANY_TO_ONE
     _class_descriptor = "assets → zones"
-    _component_type_from = "assets"
-    _component_type_to = "zones"
+    component_type_from_: str = "assets"
+    component_type_to_: str = "zones"
 
     ###################
     # INSTANCE FIELDS #
@@ -519,25 +601,16 @@ class AssetToZone(Linkage):
     # None
 
 
-class CandidateFuelToFinalFuel(Linkage):
-    ####################
-    # CLASS ATTRIBUTES #
-    ####################
+class AssetToCaisoTxConstraint(Linkage):
     _RELATIONSHIP_TYPE = LinkageRelationshipType.MANY_TO_MANY
-    _class_descriptor = "candidate fuels → final fuels"
-    _component_type_from = "candidate_fuels"
-    _component_type_to = "final_fuels"
-    _attribute_file = "candidate_fuels_to_final_fuels.csv"
+    _class_descriptor = "assets → CAISO tx constraint"
+    component_type_from_: str = "assets"
+    component_type_to_: str = "caiso_tx_constraints"
+    SAVE_PATH = "assets_to_caiso_tx_constraints.csv"
 
-    ###################
-    # INSTANCE FIELDS #
-    ###################
-    blend_limit_fraction: Optional[ts.FractionalTimeseries] = pydantic.Field(
-        default_freq="YS",
-        up_method="interpolate",
-        down_method="annual",
-    )
-
+    hsn_factor: float
+    ssn_factor: float
+    eods_factor: float
 
 class CandidateFuelToPollutant(Linkage):
     ####################
@@ -545,44 +618,93 @@ class CandidateFuelToPollutant(Linkage):
     ####################
     _RELATIONSHIP_TYPE = LinkageRelationshipType.MANY_TO_MANY
     _class_descriptor = "candidate fuels → pollutants"
-    _component_type_from = "candidate_fuels"
-    _component_type_to = "pollutants"
-    _attribute_file = "candidate_fuels_to_pollutants.csv"
+    component_type_from_: str = "candidate_fuels"
+    component_type_to_: str = "pollutants"
+    SAVE_PATH = "candidate_fuels_to_pollutants.csv"
 
     ###################
     # INSTANCE FIELDS #
     ###################
-    net_emission_factor: typing.Optional[ts.NumericTimeseries] = pydantic.Field(
+    net_emission_factor: Optional[ts.NumericTimeseries] = pydantic.Field(
         default_freq="YS",
         up_method="interpolate",
         down_method="annual",
         units=get_units("emission_factor"),
     )
 
-    gross_emission_factor: typing.Optional[ts.NumericTimeseries] = pydantic.Field(
+    gross_emission_factor: Optional[ts.NumericTimeseries] = pydantic.Field(
         default_freq="YS",
         up_method="interpolate",
         down_method="annual",
         units=get_units("emission_factor"),
     )
 
-    upstream_emission_factor: typing.Optional[ts.NumericTimeseries] = pydantic.Field(
+    upstream_emission_factor: Optional[ts.NumericTimeseries] = pydantic.Field(
         default_freq="YS",
         up_method="interpolate",
         down_method="annual",
         units=get_units("emission_factor"),
     )
 
-    gross_emissions_trajectory_override: typing.Optional[ts.NumericTimeseries] = pydantic.Field(
+    gross_emissions_trajectory_override: Optional[ts.NumericTimeseries] = pydantic.Field(
         None, default_freq="YS", up_method="interpolate", down_method="annual", units=get_units
     )
 
-    net_emissions_trajectory_override: typing.Optional[ts.NumericTimeseries] = pydantic.Field(
+    net_emissions_trajectory_override: Optional[ts.NumericTimeseries] = pydantic.Field(
         None, default_freq="YS", up_method="interpolate", down_method="annual", units=get_units
     )
 
-    upstream_emissions_trajectory_override: typing.Optional[ts.NumericTimeseries] = pydantic.Field(
+    upstream_emissions_trajectory_override: Optional[ts.NumericTimeseries] = pydantic.Field(
         None, default_freq="YS", up_method="interpolate", down_method="annual", units=get_units
+    )
+
+    # attributes that will be set during stock rollover calculations
+    out_net_emissions: ts.NumericTimeseries = pydantic.Field(
+        default_factory=lambda: ts.NumericTimeseries.default_factory(value=np.nan),
+        default_freq="YS",
+        up_method="interpolate",
+        down_method="annual",
+        weather_year=False,
+    )
+
+    out_gross_emissions: ts.NumericTimeseries = pydantic.Field(
+        default_factory=lambda: ts.NumericTimeseries.default_factory(value=np.nan),
+        default_freq="YS",
+        up_method="interpolate",
+        down_method="annual",
+        weather_year=False,
+    )
+
+    out_upstream_emissions: ts.NumericTimeseries = pydantic.Field(
+        default_factory=lambda: ts.NumericTimeseries.default_factory(value=np.nan),
+        default_freq="YS",
+        up_method="interpolate",
+        down_method="annual",
+        weather_year=False,
+    )
+
+    out_net_emissions_CO2e: ts.NumericTimeseries = pydantic.Field(
+        default_factory=lambda: ts.NumericTimeseries.default_factory(value=np.nan),
+        default_freq="YS",
+        up_method="interpolate",
+        down_method="annual",
+        weather_year=False,
+    )
+
+    out_gross_emissions_CO2e: ts.NumericTimeseries = pydantic.Field(
+        default_factory=lambda: ts.NumericTimeseries.default_factory(value=np.nan),
+        default_freq="YS",
+        up_method="interpolate",
+        down_method="annual",
+        weather_year=False,
+    )
+
+    out_upstream_emissions_CO2e: ts.NumericTimeseries = pydantic.Field(
+        default_factory=lambda: ts.NumericTimeseries.default_factory(value=np.nan),
+        default_freq="YS",
+        up_method="interpolate",
+        down_method="annual",
+        weather_year=False,
     )
 
 
@@ -592,9 +714,9 @@ class BiomassResourceToCandidateFuel(Linkage):
     ####################
     _RELATIONSHIP_TYPE = LinkageRelationshipType.MANY_TO_ONE
     _class_descriptor = "biomass resources → candidate fuels"
-    _component_type_from = "biomass_resources"
-    _component_type_to = "candidate_fuels"
-    _attribute_file = "biomass_resources_to_candidate_fuels.csv"
+    component_type_from_: str = "biomass_resources"
+    component_type_to_: str = "candidate_fuels"
+    SAVE_PATH = "biomass_resources_to_candidate_fuels.csv"
 
     ###################
     # INSTANCE FIELDS #
@@ -614,8 +736,8 @@ class ElectrolyzerToCandidateFuel(Linkage):
     ####################
     _RELATIONSHIP_TYPE = LinkageRelationshipType.MANY_TO_ONE
     _class_descriptor = "electrolyzers → candidate fuels"
-    _component_type_from = "electrolyzers"
-    _component_type_to = "candidate_fuels"
+    component_type_from_: str = "electrolyzers"
+    component_type_to_: str = "candidate_fuels"
 
 
 class FuelZoneToCandidateFuel(Linkage):
@@ -624,12 +746,20 @@ class FuelZoneToCandidateFuel(Linkage):
     ####################
     _RELATIONSHIP_TYPE = LinkageRelationshipType.MANY_TO_ONE
     _class_descriptor = "fuel zones → candidate fuels"
-    _component_type_from = "fuel_zones"
-    _component_type_to = "candidate_fuels"
+    component_type_from_: str = "fuel_zones"
+    component_type_to_: str = "candidate_fuels"
 
     ###################
     # INSTANCE FIELDS #
     ###################
+
+    @property
+    def candidate_fuel(self):
+        return self.instance_to
+
+    @property
+    def fuel_zone(self):
+        return self.instance_from
 
 
 class ElectrolyzerToFuelZone(Linkage):
@@ -638,8 +768,8 @@ class ElectrolyzerToFuelZone(Linkage):
     ####################
     _RELATIONSHIP_TYPE = LinkageRelationshipType.MANY_TO_ONE
     _class_descriptor = "electrolyzers → fuel zones"
-    _component_type_from = "electrolyzers"
-    _component_type_to = "fuel_zones"
+    component_type_from_: str = "electrolyzers"
+    component_type_to_: str = "fuel_zones"
 
 
 class ElectrolyzerToZone(Linkage):
@@ -648,8 +778,8 @@ class ElectrolyzerToZone(Linkage):
     ####################
     _RELATIONSHIP_TYPE = LinkageRelationshipType.MANY_TO_ONE
     _class_descriptor = "electrolyzers → zones"
-    _component_type_from = "electrolyzers"
-    _component_type_to = "zones"
+    component_type_from_: str = "electrolyzers"
+    component_type_to_: str = "zones"
 
     ###################
     # INSTANCE FIELDS #
@@ -663,8 +793,8 @@ class FuelStorageToFuelZone(Linkage):
     ####################
     _RELATIONSHIP_TYPE = LinkageRelationshipType.MANY_TO_ONE
     _class_descriptor = "fuel storages → fuel zones"
-    _component_type_from = "fuel_storages"
-    _component_type_to = "fuel_zones"
+    component_type_from_: str = "fuel_storages"
+    component_type_to_: str = "fuel_zones"
 
     ###################
     # INSTANCE FIELDS #
@@ -678,8 +808,8 @@ class FuelStorageToCandidateFuel(Linkage):
     ####################
     _RELATIONSHIP_TYPE = LinkageRelationshipType.MANY_TO_ONE
     _class_descriptor = "fuel storages → candidate fuels"
-    _component_type_from = "fuel_storages"
-    _component_type_to = "candidate_fuels"
+    component_type_from_: str = "fuel_storages"
+    component_type_to_: str = "candidate_fuels"
 
     ###################
     # INSTANCE FIELDS #
@@ -693,8 +823,8 @@ class FuelStorageToZone(Linkage):
     ####################
     _RELATIONSHIP_TYPE = LinkageRelationshipType.MANY_TO_ONE
     _class_descriptor = "fuel storages → zones"
-    _component_type_from = "fuel_storages"
-    _component_type_to = "zones"
+    component_type_from_: str = "fuel_storages"
+    component_type_to_: str = "zones"
 
     ###################
     # INSTANCE FIELDS #
@@ -708,8 +838,8 @@ class FuelConversionPlantToFuelZone(Linkage):
     ####################
     _RELATIONSHIP_TYPE = LinkageRelationshipType.MANY_TO_ONE
     _class_descriptor = "fuel conversion plants → fuel zones"
-    _component_type_from = "fuel_conversion_plants"
-    _component_type_to = "fuel_zones"
+    component_type_from_: str = "fuel_conversion_plants"
+    component_type_to_: str = "fuel_zones"
 
     ###################
     # INSTANCE FIELDS #
@@ -723,8 +853,8 @@ class FuelConversionPlantToCandidateFuel(Linkage):
     ####################
     _RELATIONSHIP_TYPE = LinkageRelationshipType.MANY_TO_MANY
     _class_descriptor = "fuel conversion plants → candidate fuels"
-    _component_type_from = "fuel_conversion_plants"
-    _component_type_to = "candidate_fuels"
+    component_type_from_: str = "fuel_conversion_plants"
+    component_type_to_: str = "candidate_fuels"
 
     ###################
     # INSTANCE FIELDS #
@@ -738,8 +868,8 @@ class FuelConversionPlantToZone(Linkage):
     ####################
     _RELATIONSHIP_TYPE = LinkageRelationshipType.MANY_TO_ONE
     _class_descriptor = "fuel conversion plants → zones"
-    _component_type_from = "fuel_conversion_plants"
-    _component_type_to = "zones"
+    component_type_from_: str = "fuel_conversion_plants"
+    component_type_to_: str = "zones"
 
     ###################
     # INSTANCE FIELDS #
@@ -753,8 +883,8 @@ class FinalFuelToFuelZone(Linkage):
     ####################
     _RELATIONSHIP_TYPE = LinkageRelationshipType.MANY_TO_ONE
     _class_descriptor = "final fuels → fuel zones"
-    _component_type_from = "final_fuels"
-    _component_type_to = "fuel_zones"
+    component_type_from_: str = "final_fuels"
+    component_type_to_: str = "fuel_zones"
 
     ###################
     # INSTANCE FIELDS #
@@ -768,9 +898,9 @@ class ZoneToTransmissionPath(Linkage):
     ####################
     _RELATIONSHIP_TYPE = LinkageRelationshipType.MANY_TO_MANY
     _class_descriptor = "zones → transmission paths"
-    _component_type_from = "zones"
-    _component_type_to = "tx_paths"
-    _attribute_file = "zones_to_tx_paths.csv"
+    component_type_from_: str = "zones"
+    component_type_to_: str = "tx_paths"
+    SAVE_PATH = "zones_to_tx_paths.csv"
 
     ###################
     # INSTANCE FIELDS #
@@ -781,7 +911,7 @@ class ZoneToTransmissionPath(Linkage):
     from_zone: bool = False
     to_zone: bool = False
 
-    @pydantic.root_validator
+    @pydantic.root_validator(skip_on_failure=True)
     def linkage_is_from_zone_xor_to_zone(cls, values):
         """Validate that exactly one of `from_zone` and `to_zone` is set to True."""
         if not values["from_zone"] and not values["to_zone"]:
@@ -796,38 +926,48 @@ class ZoneToTransmissionPath(Linkage):
             return values
 
 
-class FuelZoneToFuelTransportation(Linkage):
-    ####################
-    # CLASS ATTRIBUTES #
-    ####################
-    _RELATIONSHIP_TYPE = LinkageRelationshipType.MANY_TO_MANY
-    _class_descriptor = "fuel zones → fuel transportations"
-    _component_type_from = "fuel_zones"
-    _component_type_to = "fuel_transportations"
-    _attribute_file = "fuel_zones_to_fuel_transportations.csv"
+class FromZoneToPath(ZoneToTransmissionPath):
+    SAVE_PATH = None
+    from_zone: bool = True
 
-    ###################
-    # INSTANCE FIELDS #
-    ###################
-    # user should specify the from and to zone of each fuel transportation
-    # TODO: do we need a validation so that there are only two zones linked to the same line?
-    # TODO: validation to make sure there is only one from_zone for each tx_path etc.
-    from_zone: bool = False
-    to_zone: bool = False
 
-    @pydantic.root_validator
-    def linkage_is_from_zone_xor_to_zone(cls, values):
-        """Validate that exactly one of `from_zone` and `to_zone` is set to True."""
-        if not values["from_zone"] and not values["to_zone"]:
-            raise ValueError(
-                f"{cls.__name__} linkage for {values['name']} must have either 'from_zone' or 'to_zone' set to True."
-            )
-        elif values["from_zone"] and values["to_zone"]:
-            raise ValueError(
-                f"{cls.__name__} linkage for {values['name']} must have either 'from_zone' or 'to_zone' set to True, but not both."
-            )
-        else:
-            return values
+class ToZoneToPath(ZoneToTransmissionPath):
+    SAVE_PATH = None
+    to_zone: bool = True
+
+
+# class FuelZoneToFuelTransportation(Linkage):
+#     ####################
+#     # CLASS ATTRIBUTES #
+#     ####################
+#     _RELATIONSHIP_TYPE = LinkageRelationshipType.MANY_TO_MANY
+#     _class_descriptor = "fuel zones → fuel transportations"
+#     _component_type_from = "fuel_zones"
+#     _component_type_to = "fuel_transportations"
+#     SAVE_PATH = "fuel_zones_to_fuel_transportations.csv"
+#
+#     ###################
+#     # INSTANCE FIELDS #
+#     ###################
+#     # user should specify the from and to zone of each fuel transportation
+#     # TODO: do we need a validation so that there are only two zones linked to the same line?
+#     # TODO: validation to make sure there is only one from_zone for each tx_path etc.
+#     from_zone: bool = False
+#     to_zone: bool = False
+#
+#     @pydantic.root_validator(skip_on_failure=True)
+#     def linkage_is_from_zone_xor_to_zone(cls, values):
+#         """Validate that exactly one of `from_zone` and `to_zone` is set to True."""
+#         if not values["from_zone"] and not values["to_zone"]:
+#             raise ValueError(
+#                 f"{cls.__name__} linkage for {values['name']} must have either 'from_zone' or 'to_zone' set to True."
+#             )
+#         elif values["from_zone"] and values["to_zone"]:
+#             raise ValueError(
+#                 f"{cls.__name__} linkage for {values['name']} must have either 'from_zone' or 'to_zone' set to True, but not both."
+#             )
+#         else:
+#             return values
 
 
 class ZoneToZone(Linkage):
@@ -836,8 +976,8 @@ class ZoneToZone(Linkage):
     ####################
     _RELATIONSHIP_TYPE = LinkageRelationshipType.MANY_TO_MANY
     _class_descriptor = "zones → zones"
-    _component_type_to = "subzones"
-    _component_type_from = "parent_zones"
+    component_type_to_: str = "subzones"
+    component_type_from_: str = "parent_zones"
 
     ###################
     # INSTANCE FIELDS #
@@ -851,8 +991,8 @@ class FuelZoneToElectricZone(Linkage):
     ####################
     _RELATIONSHIP_TYPE = LinkageRelationshipType.MANY_TO_MANY
     _class_descriptor = "fuel zones → electric zones"
-    _component_type_to = "fuel_zones"
-    _component_type_from = "zones"
+    component_type_to_: str = "fuel_zones"
+    component_type_from_: str = "zones"
 
     ###################
     # INSTANCE FIELDS #
@@ -860,63 +1000,136 @@ class FuelZoneToElectricZone(Linkage):
     # None
 
 
-class FuelTransportationToCandidateFuel(Linkage):
-    ####################
-    # CLASS ATTRIBUTES #
-    ####################
-    _RELATIONSHIP_TYPE = LinkageRelationshipType.MANY_TO_ONE
-    _class_descriptor = "fuel transportations → candidate fuels"
-    _component_type_from = "fuel_transportations"
-    _component_type_to = "candidate_fuels"
-
-    ###################
-    # INSTANCE FIELDS #
-    ###################
-    # None
+# class FuelTransportationToCandidateFuel(Linkage):
+#     ####################
+#     # CLASS ATTRIBUTES #
+#     ####################
+#     _RELATIONSHIP_TYPE = LinkageRelationshipType.MANY_TO_ONE
+#     _class_descriptor = "fuel transportations → candidate fuels"
+#     _component_type_from = "fuel_transportations"
+#     _component_type_to = "candidate_fuels"
+#
+#     ###################
+#     # INSTANCE FIELDS #
+#     ###################
+#     # None
 
 
 # TODO: Not sure I need to create _instances class attribute every time... In general, there's a lot of boilerplate here.
 # TODO: Need to get the "from" and "to" direction consistent
 
 
-@enum.unique
-class DeliverabilityStatus(enum.Enum):
-    FULLY_DELIVERABLE = "FCDS"
-    ENERGY_ONLY = "EO"
-    BOTH = "Both"
 
-
-class AllToPolicy(Linkage):
+class _AllToPolicy(Linkage):
     ####################
     # CLASS ATTRIBUTES #
     ###################
     _RELATIONSHIP_TYPE = LinkageRelationshipType.MANY_TO_MANY
     _class_descriptor = "anything → policies"
-    _component_type_from = ""  # this will be overridden in get_dir
-    _component_type_to = "policies"
+    component_type_from_: str = ""  # this will be overridden in get_dir
+    component_type_to_: str = "policies"
     # _map_file = "all_to_policies/mapping.csv"
-    _attribute_file = "all_to_policies.csv"
+    SAVE_PATH = "all_to_policies.csv"
 
+    multiplier: Optional[ts.NumericTimeseries] = pydantic.Field(
+        None, default_freq="YS", up_method="ffill", down_method="sum", alias="nqc"
+    )
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+
+        # Check if "component_type_from_" and "component_type_to_" exist if reading from json files in PCM
+        if "component_type_from_" not in kwargs:
+            self.component_type_from_: str = self._get_component_type_from_()
+        if "component_type_to_" not in kwargs:
+            self.component_type_to_: str = self._get_component_type_to_()
+
+    def _get_component_type_from_(self):
+        from new_modeling_toolkit import system
+
+        if isinstance(self.instance_from, system.electric.load_component.Load):
+            component_type_from = "loads"
+        elif isinstance(self.instance_from, system.electric.elcc.ELCCSurface):
+            component_type_from = "elcc_surfaces"
+        elif isinstance(self.instance_from, system.fuel.candidate_fuel.CandidateFuel):
+            component_type_from = "candidate_fuels"
+        elif isinstance(self.instance_from, system.fuel.final_fuel.FinalFuel):
+            component_type_from = "final_fuels"
+        elif isinstance(self.instance_from, system.electric.reserve.Reserve):
+            component_type_from = "reserves"
+        elif isinstance(self.instance_from, system.electric.zone.Zone):
+            component_type_from = "zones"
+        elif isinstance(self.instance_from, system.electric.resources.generic.GenericResource):
+            component_type_from = "resources"
+        elif isinstance(self.instance_from, system.electric.tx_path.TxPath):
+            component_type_from = "tx_paths"
+        elif isinstance(self.instance_from, system.pollution.negative_emissions_technology.NegativeEmissionsTechnology):
+            component_type_from = "negative_emissions_technologies"
+        elif isinstance(self.instance_from, system.generics.plant.Plant):
+            component_type_from = "plants"
+        elif isinstance(self.instance_from, system.generics.demand.Demand):
+            component_type_from = "demands"
+        elif isinstance(self.instance_from, system.generics.transportation.Transportation):
+            component_type_from = "transportations"
+        elif isinstance(self.instance_from, system.pollution.pollutant.Pollutant):
+            component_type_from = "pollutants"
+        elif isinstance(self.instance_from, system.electric.CaisoTxConstraint):
+            component_type_from = "caiso_tx_constraints"
+        elif isinstance(self.instance_from, system.asset.Asset):
+            if isinstance(self.instance_to, system.policy.PlanningReserveMargin) or isinstance(
+                self.instance_to, system.policy.EnergyReserveMargin
+            ):
+                component_type_from = "assets_"
+            else:
+                component_type_from = "assets"
+        else:
+            raise ValueError(
+                f"Unrecognized type for `instance_from` in AllToPolicy linkage `{self.name}`: `{type(self.instance_from)}`"
+            )
+
+        return component_type_from
+
+    def _get_component_type_to_(self):
+        from new_modeling_toolkit.system import policy
+
+        if isinstance(self.instance_to, policy.AnnualEmissionsPolicy):
+            component_type_to = "emissions_policies"
+        elif isinstance(self.instance_to, policy.AnnualEnergyStandard):
+            component_type_to = "annual_energy_policies"
+        elif isinstance(self.instance_to, policy.HourlyEnergyStandard):
+            component_type_to = "hourly_energy_policies"
+        elif isinstance(self.instance_to, policy.PlanningReserveMargin):
+            component_type_to = "prm_policies"
+        elif isinstance(self.instance_to, policy.EnergyReserveMargin):
+            component_type_to = "erm_policies"
+            if not isinstance(self, ERMContribution):
+                raise ValueError(
+                    f"ERM policy linkages must be explicitly modeled as ERMContribution linkages."
+                    f"Change the AllToPolicy linkage `{self.name}` to an ERMContribution linkage."
+                )
+        else:
+            raise ValueError(
+                f"Unrecognized type for `instance_to` in AllToPolicy linkage `{self.name}`: `{type(self.instance_to)}`"
+            )
+
+        return component_type_to
+
+
+class AllToPolicy(_AllToPolicy):
     ###################
     # INSTANCE FIELDS #
     ###################
-    multiplier: typing.Optional[ts.NumericTimeseries] = pydantic.Field(
-        None, default_freq="YS", up_method="ffill", down_method="annual", alias="nqc"
-    )
-    # TODO: should be pint.Unit but don't know how to do it
-    multiplier_unit: typing.Optional[str] = None
-    attribute: typing.Optional[str] = None  # attribute in the RESOLVE model
-    forward_dir_multiplier: typing.Optional[ts.NumericTimeseries] = pydantic.Field(
-        None, default_freq="YS", up_method="ffill", down_method="annual"
+    attribute: Annotated[str | None, Metadata(default_exclude=True)] = None  # attribute in the RESOLVE model
+    forward_dir_multiplier: Optional[ts.NumericTimeseries] = pydantic.Field(
+        None, default_freq="YS", up_method="ffill", down_method="sum"
     )  # used when Transmission Path to Policy linkage
-    reverse_dir_multiplier: typing.Optional[ts.NumericTimeseries] = pydantic.Field(
-        None, default_freq="YS", up_method="ffill", down_method="annual"
+    reverse_dir_multiplier: Optional[ts.NumericTimeseries] = pydantic.Field(
+        None, default_freq="YS", up_method="ffill", down_method="sum"
     )  # used when Transmission Path to Policy linkage
 
-    deliverability_status: DeliverabilityStatus = DeliverabilityStatus.BOTH
+    fully_deliverable: bool = True
 
     # Used when Transmission Path to Policy Linkage
-    @pydantic.root_validator
+    @pydantic.root_validator(skip_on_failure=True)
     def validate_single_xor_bidirectional_multiplier(cls, values):
         """Validate that the linkage either has both 'forward_dir_multiplier' & 'reverse_dir_multiplier' defined,
         xor only has 'multiplier' defined. The two types of multipliers shouldn't exist at the same time."""
@@ -933,7 +1146,68 @@ class AllToPolicy(Linkage):
             return values
 
 
-class FinelFuelToAnnualEmissionsPolicy(AllToPolicy):
+# Creating aliases for convenience
+class EmissionsContribution(_AllToPolicy):
+    SAVE_PATH = "resource_fuel_emissions_contributions.csv"
+    multiplier: Annotated[
+        ts.NumericTimeseries | None, Metadata(excel_short_title="tonne/MWh", units=units.tonne / units.MWh)
+    ] = pydantic.Field(None, default_freq="YS", up_method="ffill", down_method="mean", title="Emissions Rate")
+
+
+class TxEmissionsContribution(EmissionsContribution):
+    SAVE_PATH = "tx_emissions_contributions.csv"
+    forward_dir_multiplier: Annotated[
+        ts.NumericTimeseries | None, Metadata(excel_short_title="Rate", units=units.tonne / units.MWh)
+    ] = pydantic.Field(
+        None, default_freq="YS", up_method="ffill", down_method="sum", title="Forward Emissions Rate (per-MWh)"
+    )
+    reverse_dir_multiplier: Annotated[
+        ts.NumericTimeseries | None, Metadata(excel_short_title="Rate", units=units.tonne / units.MWh)
+    ] = pydantic.Field(
+        None, default_freq="YS", up_method="ffill", down_method="sum", title="Reverse Emissions Rate (per-MWh)"
+    )
+
+
+# TODO: Add TransportationEmissionsContribution
+
+
+class AnnualEnergyStandardContribution(_AllToPolicy):
+    SAVE_PATH = "annual_energy_standard_contributions.csv"
+    multiplier: Annotated[ts.NumericTimeseries | None, Metadata(excel_short_title="%")] = pydantic.Field(
+        None, default_freq="YS", up_method="ffill", down_method="annual", title="Qualifying Generation (%)"
+    )
+
+
+class HourlyEnergyStandardContribution(_AllToPolicy):
+    SAVE_PATH = "hourly_energy_standard_contributions.csv"
+    multiplier: Annotated[ts.NumericTimeseries | None, Metadata(excel_short_title="%")] = pydantic.Field(
+        None, default_freq="YS", up_method="ffill", down_method="annual", title="Qualifying Generation (%)"
+    )
+
+
+class ERMContribution(_AllToPolicy):
+    SAVE_PATH = "erm_contributions.csv"
+    multiplier: Annotated[ts.NumericTimeseries | None, Metadata(excel_short_title="%")] = pydantic.Field(
+        default_factory=ts.Timeseries.one,
+        default_freq="H",
+        up_method="ffill",
+        down_method="mean",
+        weather_year=True,
+        title="ERM Hourly Fraction",
+    )
+
+class ReliabilityContribution(_AllToPolicy):
+    SAVE_PATH = "reliability_contributions.csv"
+    attribute: Annotated[str | None, Metadata(default_exclude=True)] = None  # attribute in the RESOLVE model
+    fully_deliverable: bool = True
+    multiplier: Annotated[ts.NumericTimeseries | None, Metadata(excel_short_title="%")] = pydantic.Field(
+        None, default_freq="YS", up_method="ffill", down_method="annual", title="Qualifying Capacity"
+    )
+
+class ELCCReliabilityContribution(_AllToPolicy):
+    SAVE_PATH = "elcc_reliability_contributions.csv"
+
+class FinalFuelToAnnualEmissionsPolicy(AllToPolicy):
     """Temporary subclass only so that final fuels are correctly linked to annual emissions policies."""
 
     ####################
@@ -941,13 +1215,19 @@ class FinelFuelToAnnualEmissionsPolicy(AllToPolicy):
     ###################
     _RELATIONSHIP_TYPE = LinkageRelationshipType.MANY_TO_MANY
     _class_descriptor = "final fuel → annual emissions policy"
-    _component_type_from = "final_fuels"
-    _component_type_to = "policies"
+    component_type_from_: str = "final_fuels"
+    component_type_to_: str = "policies"
 
     ###################
     # INSTANCE FIELDS #
     ###################
     # None
+
+
+class ELCCFacetToSurface(Linkage):
+    _RELATIONSHIP_TYPE = LinkageRelationshipType.MANY_TO_ONE
+    component_type_from_: str = "facets"
+    component_type_to_: str = "surface"
 
 
 class AssetToELCC(Linkage):
@@ -956,23 +1236,21 @@ class AssetToELCC(Linkage):
     ####################
     _RELATIONSHIP_TYPE = LinkageRelationshipType.MANY_TO_MANY
     _class_descriptor = "assets → ELCC"
-    _component_type_from = "assets"
-    _component_type_to = "elcc_surfaces"
-    _attribute_file = "assets_to_elcc.csv"
+    component_type_from_: str = "assets"
+    component_type_to_: str = "elcc_surfaces"
+    SAVE_PATH = "assets_to_elcc.csv"
 
     ###################
     # INSTANCE FIELDS #
     ###################
-    multiplier: typing.Optional[float] = None
-    multiplier_unit: typing.Optional[str] = None
-    attribute: typing.Literal["power", "energy"] = "power"
-    elcc_axis_index: typing.Optional[int] = None
-    elcc_axis_multiplier: typing.Optional[float] = None
+    # TODO 2024-08-15: As far as I can tell, `multiplier` and `multiplier_unit` don't do anything?
+    multiplier: Annotated[float | None, Metadata(default_exclude=True)] = None
+    multiplier_unit: Annotated[str | None, Metadata(default_exclude=True)] = None
+    attribute: Annotated[Literal["power", "energy"], Metadata(default_exclude=True)] = "power"
+    elcc_axis_index: int = 1
+    elcc_axis_multiplier: float = 1
 
-    # Convert strings that look like floats to integers for integer fields
-    _convert_int = pydantic.validator("elcc_axis_index", allow_reuse=True, pre=True)(convert_str_float_to_int)
-
-    @pydantic.root_validator
+    @pydantic.root_validator(skip_on_failure=True)
     def has_axis_index_if_multiplier(cls, values):
         if values["elcc_axis_multiplier"] is not None:
             assert values["elcc_axis_index"] is not None, (
@@ -983,195 +1261,78 @@ class AssetToELCC(Linkage):
         return values
 
 
-def find_component_type(components_dict: dict[str, dict[str, component.Component]], instance_name: str):
-    """Try to find the component with the given instance_name.
-
-    Args:
-        components_dict: Nested dictionary of all System instance components.
-            First level of the dict is the name of the component type (class name),
-            while dict nested inside holds the various instances of that component type.
-        inst_name: The name of the current instance being queried.
-
-    Returns:
-        The name of the component type of the current instance under query.
-
-    """
-    matching_component_types = [
-        component_type
-        for component_type, component_dict_one_type in components_dict.items()
-        if instance_name in component_dict_one_type.keys()
-    ]
-
-    # Warn if no components found or multiple components found (i.e., components with the same name but different types)
-    if not matching_component_types:
-        err = f"{instance_name} not found in System components"
-        logger.exception(err)
-        return None
-    elif (
-        len(matching_component_types) > 1
-        and set(matching_component_types) != {"assets", "resources", "plants"}
-        and set(matching_component_types) != {"assets", "generic_assets"}
-    ):
-        logger.debug(
-            f"Multiple System components found with name '{instance_name}' in: {matching_component_types}. Using {matching_component_types[0]}"
-        )
-
-    # Hacky way to return `assets` as component type as a last resort
-    if "resources" in matching_component_types:
-        return "resources"
-    if "plants" in matching_component_types:
-        return "plants"
-    if "tx_paths" in matching_component_types:
-        return "tx_paths"
-    if "assets" in matching_component_types:
-        return "assets"
-    else:
-        return matching_component_types[0]
-
-
-class DeviceToFinalFuel(Linkage):
+class XToFinalFuel(Linkage):
     ####################
     # CLASS ATTRIBUTES #
     ####################
+    _RELATIONSHIP_TYPE = LinkageRelationshipType.ONE_TO_MANY
 
-    _class_descriptor = "devices → final fuels"
-    _component_type_from = "devices"
-    _component_type_to = "final_fuels"
-    _attribute_file = "devices_to_final_fuels.csv"
-
-    ###################
-    # INSTANCE FIELDS #
-    ###################
-    fuel_share_of_service_demand: typing.Optional[str] = None
-
-    device_efficiency: ts.NumericTimeseries = pydantic.Field(
-        default_freq="YS", up_method="interpolate", down_method="annual", units=get_units("device_efficiency")
-    )
+    _class_descriptor = "component → final fuels"
+    component_type_from_: str = "component"
+    component_type_to_: str = "final_fuels"
+    SAVE_PATH = "component_to_final_fuels.csv"
 
     # attributes that will be set during stock rollover calculations
-    out_final_energy_demand: typing.Optional[ts.NumericTimeseries] = pydantic.Field(
-        default_freq="YS", up_method="interpolate", down_method="annual", units=get_units("out_final_energy_demand")
-    )
-
-
-class StockRolloverSubsectorToDevice(Linkage):
-    ####################
-    # CLASS ATTRIBUTES #
-    ####################
-
-    _class_descriptor = "stock rollover subsectors → devices"
-    _component_type_from = "stock_rollover_subsectors"
-    _component_type_to = "devices"
-    _attribute_file = "stock_rollover_subsectors_to_devices.csv"
-
-    ###################
-    # INSTANCE FIELDS #
-    ###################
-
-
-class BuildingShellSubsectorToBuildingShellType(Linkage):
-    ####################
-    # CLASS ATTRIBUTES #
-    ####################
-
-    _class_descriptor = "building shell subsectors → building shell types"
-    _component_type_from = "building_shell_subsectors"
-    _component_type_to = "building_shell_types"
-    _attribute_file = "building_shell_subsectors_to_building_shell_types.csv"
-
-    ###################
-    # INSTANCE FIELDS #
-    ###################
-
-
-class BuildingShellSubsectorToSector(Linkage):
-    ####################
-    # CLASS ATTRIBUTES #
-    ####################
-
-    _class_descriptor = "building shell subsectors → sectors"
-    _component_type_from = "building_shell_subsectors"
-    _component_type_to = "sectors"
-
-
-class EnergyDemandSubsectorToFinalFuel(Linkage):
-    """
-    Defines the linkage betweeen energy demand subsectors and final fuels. It is used to define the assumed growth
-    of fuel demands within a subsector and the extent and cost of efficiency by fuel within a subsector. Note efficiency
-    is applied after fuel conversions.
-    """
-
-    ####################
-    # CLASS ATTRIBUTES #
-    ####################
-
-    _class_descriptor = "energy demand subsectors → final fuels"
-    _component_type_from = "energy_demand_subsectors"
-    _component_type_to = "final_fuels"
-    _attribute_file = "energy_demand_subsectors_to_final_fuels.csv"
-
-    ###################
-    # INSTANCE FIELDS #
-    ###################
-
-    default_energy_demand_trajectory: typing.Optional[ts.NumericTimeseries] = pydantic.Field(
+    out_energy_demand: ts.NumericTimeseries = pydantic.Field(
+        default_factory=lambda: ts.NumericTimeseries.default_factory(value=np.nan),
         default_freq="YS",
         up_method="interpolate",
-        down_method="annual",
-        units=get_units("default_energy_demand_trajectory"),
-    )
-    efficiency_fraction: typing.Optional[ts.FractionalTimeseries] = pydantic.Field(
-        default_freq="YS", up_method="interpolate", down_method="annual", units=get_units("efficiency_fraction")
-    )
-    efficiency_cost: typing.Optional[ts.NumericTimeseries] = pydantic.Field(
-        default_freq="YS", up_method="interpolate", down_method="annual", units=get_units("efficiency_cost")
-    )
-
-    # attributes that will be set during stock rollover calculations
-    out_final_energy_demand: typing.Optional[ts.NumericTimeseries] = pydantic.Field(
-        default_freq="YS", up_method="interpolate", down_method="annual", units=get_units("out_final_energy_demand")
-    )
-    out_efficiency_cost: typing.Optional[ts.NumericTimeseries] = pydantic.Field(
-        default_freq="YS", up_method="interpolate", down_method="annual", units=get_units("out_efficiency_cost")
+        down_method="first",
+        weather_year=True,
     )
 
 
-class NonEnergySubsectorToPollutant(Linkage):
-    """
-    Defines the linkage betweeen non energy subsectors and pollutants.
-    """
-
+class CandidateFuelToFinalFuel(XToFinalFuel):
     ####################
     # CLASS ATTRIBUTES #
     ####################
-
-    _class_descriptor = "non-energy subsectors → pollutants"
-    _component_type_from = "non_energy_subsectors"
-    _component_type_to = "pollutants"
-    _attribute_file = "non_energy_subsectors_to_pollutants.csv"
+    _RELATIONSHIP_TYPE = LinkageRelationshipType.ONE_TO_ONE
+    _class_descriptor = "candidate fuels → final fuels"
+    component_type_from_: str = "candidate_fuels"
+    component_type_to_: str = "final_fuels"
+    SAVE_PATH = "candidate_fuels_to_final_fuels.csv"
 
     ###################
     # INSTANCE FIELDS #
     ###################
-
-    emissions: ts.NumericTimeseries = pydantic.Field(
-        default_freq="YS", up_method="interpolate", down_method="annual", units=get_units("emissions")
-    )
-    cost: Optional[ts.NumericTimeseries] = pydantic.Field(
-        None, default_freq="YS", up_method="interpolate", down_method="annual", units=get_units("cost")
+    blend_limit_fraction: Optional[ts.NumericTimeseries] = pydantic.Field(
+        default_freq="YS", up_method="interpolate", down_method="annual", weather_year=False
     )
 
+    ######################
+    # Calculated Outputs #
+    ######################
+    out_fuel_cost: ts.NumericTimeseries = pydantic.Field(
+        default_factory=lambda: ts.NumericTimeseries.default_factory(value=np.nan),
+        default_freq="YS",
+        up_method="interpolate",
+        down_method="first",
+        weather_year=True,
+    )
 
-class NegativeEmissionsTechnologyToPollutant(Linkage):
-    """Defines the linkage between NETs and pollutants."""
+    out_net_emissions: ts.NumericTimeseries = pydantic.Field(
+        default_factory=lambda: ts.NumericTimeseries.default_factory(value=np.nan),
+        up_method="ffill",
+        down_method="annual",
+        default_freq="YS",
+        weather_year=True,
+    )
 
-    ####################
-    # CLASS ATTRIBUTES #
-    ####################
+    out_gross_emissions: ts.NumericTimeseries = pydantic.Field(
+        default_factory=lambda: ts.NumericTimeseries.default_factory(value=np.nan),
+        up_method="ffill",
+        down_method="annual",
+        default_freq="YS",
+        weather_year=True,
+    )
 
-    _class_descriptor = "negative emissions technologies → pollutants"
-    _component_type_from = "negative_emissions_technologies"
-    _component_type_to = "pollutants"
+    out_upstream_emissions: ts.NumericTimeseries = pydantic.Field(
+        default_factory=lambda: ts.NumericTimeseries.default_factory(value=np.nan),
+        up_method="ffill",
+        down_method="annual",
+        default_freq="YS",
+        weather_year=True,
+    )
 
 
 class ResourceToPollutant(Linkage):
@@ -1180,14 +1341,16 @@ class ResourceToPollutant(Linkage):
     pollutant level rather than the resource to fuel to pollutant level.
     """
 
+    _RELATIONSHIP_TYPE = LinkageRelationshipType.MANY_TO_MANY
+
     ####################
     # CLASS ATTRIBUTES #
     ####################
 
     _class_descriptor = "resources → pollutants"
-    _component_type_from = "resources"
-    _component_type_to = "pollutants"
-    _attribute_file = "resources_to_pollutants.csv"
+    component_type_from_: str = "resources"
+    component_type_to_: str = "pollutants"
+    SAVE_PATH = "resources_to_pollutants.csv"
 
     emission_factor: ts.NumericTimeseries = pydantic.Field(
         None, default_freq="YS", up_method="interpolate", down_method="annual", units=get_units
@@ -1199,14 +1362,16 @@ class TransmissionPathToPollutant(Linkage):
     Defines the linkage between transmission paths and pollutants. Used to set emission factors for Tx lines.
     """
 
+    _RELATIONSHIP_TYPE = LinkageRelationshipType.MANY_TO_MANY
+
     ####################
     # CLASS ATTRIBUTES #
     ####################
 
     _class_descriptor = "transmission paths → pollutants"
-    _component_type_from = "tx_paths"
-    _component_type_to = "pollutants"
-    _attribute_file = "tx_paths_to_pollutants.csv"
+    component_type_from_: str = "tx_paths"
+    component_type_to_: str = "pollutants"
+    SAVE_PATH = "tx_paths_to_pollutants.csv"
 
     forward_dir_multiplier: ts.NumericTimeseries = pydantic.Field(
         None, default_freq="YS", up_method="interpolate", down_method="annual", units=get_units
@@ -1214,224 +1379,3 @@ class TransmissionPathToPollutant(Linkage):
     reverse_dir_multiplier: ts.NumericTimeseries = pydantic.Field(
         None, default_freq="YS", up_method="interpolate", down_method="annual", units=get_units
     )
-
-
-class DeviceToLoad(Linkage):
-    """
-    Defines the mapping from a device to a load component. Used to populate load component annual energy with
-    outputs of stock rollover.
-    """
-
-    ####################
-    # CLASS ATTRIBUTES #
-    ####################
-
-    _class_descriptor = "devices → loads"
-    _component_type_from = "devices"
-    _component_type_to = "loads"
-    _attribute_file = "devices_to_loads.csv"
-
-    ###################
-    # INSTANCE FIELDS #
-    ###################
-
-    # The multiplier is what energy demands from devices are multiplied by to add to the annual_energy_forecast in the
-    # linked load component, if applicable. Used for unit conversion. Default value is conversion from MMBTU to MWh.
-    multiplier: float = 0.293071
-
-
-class EnergyDemandSubsectorToLoad(Linkage):
-    """
-    Defines the mapping from a energy demand subsector to a load component. Used to populate load component
-    annual energy with outputs of energy demand calculations.
-    """
-
-    ####################
-    # CLASS ATTRIBUTES #
-    ####################
-
-    _class_descriptor = "energy_demand_subsectors → loads"
-    _component_type_from = "energy_demand_subsectors"
-    _component_type_to = "loads"
-    _attribute_file = "energy_demand_subsectors_to_loads.csv"
-
-    ###################
-    # INSTANCE FIELDS #
-    ###################
-
-    # The multiplier is what energy demands from energy demand subsectors are multiplied by to add to the
-    # annual_energy_forecast in the linked load component, if applicable. Used for unit conversion.
-    # Default value is conversion from MMBTU to MWh.
-    multiplier: float = 0.293071
-
-
-class StockRolloverSubsectorToZone(Linkage):
-    ####################
-    # CLASS ATTRIBUTES #
-    ####################
-
-    _class_descriptor = "stock rollover subsectors → zones"
-    _component_type_from = "stock_rollover_subsectors"
-    _component_type_to = "zones"
-
-
-class EnergyDemandSubsectorToZone(Linkage):
-    ####################
-    # CLASS ATTRIBUTES #
-    ####################
-
-    _class_descriptor = "energy demand subsectors → zones"
-    _component_type_from = "energy_demand_subsectors"
-    _component_type_to = "zones"
-
-
-class NonEnergySubsectorToZone(Linkage):
-    ####################
-    # CLASS ATTRIBUTES #
-    ####################
-
-    _class_descriptor = "non-energy subsectors → zones"
-    _component_type_from = "non_energy_subsectors"
-    _component_type_to = "zones"
-
-
-class StockRolloverSubsectorToSector(Linkage):
-    ####################
-    # CLASS ATTRIBUTES #
-    ####################
-
-    _class_descriptor = "stock rollover subsectors → sectors"
-    _component_type_from = "stock_rollover_subsectors"
-    _component_type_to = "sectors"
-
-
-class EnergyDemandSubsectorToSector(Linkage):
-    ####################
-    # CLASS ATTRIBUTES #
-    ####################
-
-    _class_descriptor = "energy demand subsectors → sectors"
-    _component_type_from = "energy_demand_subsectors"
-    _component_type_to = "sectors"
-
-
-class NonEnergySubsectorToSector(Linkage):
-    ####################
-    # CLASS ATTRIBUTES #
-    ####################
-
-    _class_descriptor = "non-energy subsectors → sectors"
-    _component_type_from = "non_energy_subsectors"
-    _component_type_to = "sectors"
-
-
-class CCSPlantToFinalFuel(Linkage):
-    ####################
-    # CLASS ATTRIBUTES #
-    ####################
-
-    _class_descriptor = "ccs_plants → final_fuels"
-    _component_type_from = "ccs_plants"
-    _component_type_to = "final_fuels"
-    _attribute_file = "ccs_plants_to_final_fuels.csv"
-
-    ###################
-    # INSTANCE FIELDS #
-    ###################
-
-    ccs_energy_demand: typing.Optional[ts.NumericTimeseries] = pydantic.Field(
-        default_freq="YS",
-        up_method="interpolate",
-        down_method="annual",
-        description="Energy demand required to run CCS (in units of MMBtu per metric ton)",
-    )
-
-    # attributes that will be set during calculations
-    out_final_energy_demand: typing.Optional[ts.NumericTimeseries] = pydantic.Field(
-        default_freq="YS", up_method="interpolate", down_method="annual", units=get_units("out_final_energy_demand")
-    )
-
-
-class NegativeEmissionsTechnologyToFinalFuel(Linkage):
-    ####################
-    # CLASS ATTRIBUTES #
-    ####################
-
-    _class_descriptor = "negative_emissions_technologies → final_fuels"
-    _component_type_from = "negative_emissions_technologies"
-    _component_type_to = "final_fuels"
-    _attribute_file = "negative_emissions_technologies_to_final_fuels.csv"
-
-    ###################
-    # INSTANCE FIELDS #
-    ###################
-
-    energy_demand: typing.Optional[ts.NumericTimeseries] = pydantic.Field(
-        default_freq="YS",
-        up_method="interpolate",
-        down_method="annual",
-        description="Energy demand required to run NET (in units of MMBtu per metric ton)",
-    )
-
-    # attributes that will be set during calculations
-    out_final_energy_demand: typing.Optional[ts.NumericTimeseries] = pydantic.Field(
-        default_freq="YS", up_method="interpolate", down_method="annual", units=get_units("out_final_energy_demand")
-    )
-
-
-class CCSPlantToPollutant(Linkage):
-    ####################
-    # CLASS ATTRIBUTES #
-    ####################
-
-    _class_descriptor = "ccs_plants → pollutants"
-    _component_type_from = "ccs_plants"
-    _component_type_to = "pollutants"
-
-
-class NonEnergySubsectorToCCSPlant(Linkage):
-    ####################
-    # CLASS ATTRIBUTES #
-    ####################
-
-    _class_descriptor = "non_energy_subsectors → ccs_plants"
-    _component_type_from = "non_energy_subsectors"
-    _component_type_to = "ccs_plants"
-    _attribute_file = "non_energy_subsectors_to_ccs_plants.csv"
-
-    ###################
-    # INSTANCE FIELDS #
-    ###################
-
-    ccs_application_percentage: typing.Optional[ts.FractionalTimeseries] = pydantic.Field(
-        None,
-        default_freq="YS",
-        up_method="interpolate",
-        down_method="annual",
-        description="percentage of gross emissions to capture",
-    )
-
-
-class NonEnergySubsectorToEnergyDemandSubsector(Linkage):
-    ####################
-    # CLASS ATTRIBUTES #
-    ####################
-
-    _class_descriptor = "non_energy_subsectors → energy_demand_subsectors"
-    _component_type_from = "non_energy_subsectors"
-    _component_type_to = "energy_demand_subsectors"
-    _attribute_file = "non_energy_subsectors_to_energy_demand_subsectors.csv"
-
-
-if __name__ == "__main__":
-    from new_modeling_toolkit.core import dir_str, energy_demand_subsector, fuel
-
-    energy_demand_subsectors = energy_demand_subsector.EnergyDemandSubsector.from_dir(
-        dir_str.data_interim_dir / "energy_demand_subsectors"
-    )
-    fuels = fuel.FinalFuel.from_dir(dir_str.data_interim_dir / "final_fuels")
-
-    eds = energy_demand_subsectors["CA_Industry_Agriculture"]
-    f = fuels["Diesel"]
-
-    linkage = EnergyDemandSubsectorToFinalFuel(name="test linkage", _instance_from=eds, _instance_to=f)
