@@ -1,5 +1,6 @@
 from typing import Union
 
+import numpy as np
 import pandas as pd
 import pyomo.environ as pyo
 from loguru import logger
@@ -13,42 +14,33 @@ def mark_pyomo_component(func):
 
 
 def get_index_labels(model_component: Union[pyo.Param, pyo.Var, pyo.Expression, pyo.Constraint]) -> list[str]:
-    """Get the names of the indices, given a Pyomo model component instance."""
+    """
+    Get the names of the indices, given a Pyomo model component instance.
+    Unpack the tuple listed in "doc" input of Set definition
+    """
     if model_component.is_indexed():
         # If component has multiple indices, we need to do some additional unpacking using _implicit_subsets
         if model_component._implicit_subsets is not None:
-            names = [s.name for s in model_component._implicit_subsets]
+            names = [
+                elem
+                for s in model_component._implicit_subsets
+                for elem in (tuple(s.doc) if s.doc is not None else [s.name])
+            ]
+            # Several sets are multi-dimensional, which require yet another (last) unpacking
+            if "CHRONO_PERIODS_AND_TIMESTAMPS" in names:
+                tuple_pos = names.index("CHRONO_PERIODS_AND_TIMESTAMPS")
+                names[tuple_pos : tuple_pos + 1] = ("CHRONO_PERIODS", "TIMESTAMPS")
+            elif [name for name in names if name.endswith(".connect_process_outputs_index_0")]:
+                tuple_pos = names.index(
+                    [name for name in names if name.endswith(".connect_process_outputs_index_0")][0]
+                )
+                names[tuple_pos : tuple_pos + 1] = ("PROCESS_INPUT", "PROCESS_OUTPUT")
+            elif "WEATHER_PERIODS_AND_WEATHER_TIMESTAMPS" in names:
+                tuple_pos = names.index("WEATHER_PERIODS_AND_WEATHER_TIMESTAMPS")
+                names[tuple_pos : tuple_pos + 1] = ("WEATHER_PERIODS", "WEATHER_TIMESTAMPS")
+
         else:
             names = [model_component.index_set().name]
-
-        # Several sets are multi-dimensional, which require yet another (last) unpacking
-        if "TIMEPOINTS" in names:  # TIMEPOINTS are a tuple of MODEL_YEARS, REP_PERIODS, and HOURS
-            tuple_pos = names.index("TIMEPOINTS")
-            names[tuple_pos : tuple_pos + 1] = ("MODEL_YEARS", "REP_PERIODS", "HOURS")
-
-        if "MODEL_YEARS_AND_ADJACENT_REP_PERIODS" in names:
-            tuple_pos = names.index("MODEL_YEARS_AND_ADJACENT_REP_PERIODS")
-            names[tuple_pos : tuple_pos + 1] = ("MODEL_YEARS", "PREV_REP_PERIODS", "NEXT_REP_PERIODS")
-
-        if "MODEL_YEARS_AND_CHRONO_PERIODS" in names:
-            tuple_pos = names.index("MODEL_YEARS_AND_CHRONO_PERIODS")
-            names[tuple_pos : tuple_pos + 1] = ("MODEL_YEARS", "CHRONO_PERIODS")
-
-        if "DISPATCH_WINDOWS_AND_TIMESTAMPS" in names:
-            tuple_pos = names.index("DISPATCH_WINDOWS_AND_TIMESTAMPS")
-            names[tuple_pos : tuple_pos + 1] = ("DISPATCH_WINDOWS", "TIMESTAMPS")
-
-        if "RESOURCE_CANDIDATE_FUELS_FOR_EMISSIONS_POLICY" in names:
-            tuple_pos = names.index("RESOURCE_CANDIDATE_FUELS_FOR_EMISSIONS_POLICY")
-            names[tuple_pos : tuple_pos + 1] = ("EMISSIONS_POLICIES", "RESOURCES", "CANDIDATE_FUELS")
-
-        if "SIMULTANEOUS_FLOW_GROUPS_MAP" in names:
-            tuple_pos = names.index("SIMULTANEOUS_FLOW_GROUPS_MAP")
-            names[tuple_pos : tuple_pos + 1] = ("SIMULTANEOUS_FLOW_GROUPS", "TRANSMISSION_LINES")
-
-        if "PLANTS_THAT_INCREMENT_RESERVES" in names:
-            tuple_pos = names.index("PLANTS_THAT_INCREMENT_RESERVES")
-            names[tuple_pos : tuple_pos + 1] = ("PLANTS", "RESERVES")
     else:
         names = [None]
 
@@ -59,6 +51,7 @@ def convert_pyomo_object_to_dataframe(
     model_component: Union[pyo.Param, pyo.Var, pyo.Expression, pyo.Constraint],
     exception: bool = True,
     dual_only: bool = False,
+    use_doc_as_column_name: bool = False,
 ) -> pd.DataFrame:
     """Converts an object from a pyomo model (Param, Var, Expression, or Constraint) into a pandas DataFrame.
 
@@ -70,44 +63,48 @@ def convert_pyomo_object_to_dataframe(
         exception: Passthrough to `pyomo.Value()`. If True, raise an exception for uninitialized components. If False,
             return None for unintialized values.
         dual_only: for a Constraint, whether to return only the dual values
+        use_doc_as_column_name: True if the column name should be what is defined in optional `doc` attribute, otherwise the column name will be returned as the name of the component
 
     Returns:
         df: the pyomo object in DataFrame format
     """
+    column_name = [
+        model_component.name if model_component.doc is None and not use_doc_as_column_name else model_component.doc
+    ]
     if isinstance(model_component, (pyo.Param, pyo.Var, pyo.Expression)):
         # Get model component results as a dict using extract_values() method
         obj_results = {
             idx: pyo.value(v, exception=exception) if not isinstance(v, tuple) else str(v)
             for idx, v in model_component.extract_values().items()
         }
-        # TODO (2022-02-22): We could make the `get_index_labels` function do get both names and the entire index object.
         if model_component.is_indexed():
             names = get_index_labels(model_component)
-            if model_component._implicit_subsets is not None or model_component.index_set().name in [
-                "TIMEPOINTS",
-                "ADJACENT_REP_PERIODS",
-                "DISPATCH_WINDOWS_AND_TIMESTAMPS",
-                "RESOURCE_CANDIDATE_FUELS_FOR_EMISSIONS_POLICY",
-                "SIMULTANEOUS_FLOW_GROUPS_MAP",
-            ]:
-                if len(obj_results.keys()) == 0:
-                    index = pd.MultiIndex.from_tuples([(None,) * len(names)], names=names)
-                else:
-                    index = pd.MultiIndex.from_tuples(obj_results.keys(), names=names)
+            if len(names) > 1:
+                index = pd.MultiIndex.from_tuples(obj_results.keys(), names=names)
             else:
                 index = pd.Index(obj_results.keys(), name=names[0])
         else:
             # Scalar values get an empty index with no name
             index = pd.Index([None], name="[None]")
         # Create dataframe from dict
-        df = pd.DataFrame(obj_results.values(), index=index, columns=[model_component.name])
+        df = pd.DataFrame(
+            obj_results.values(),
+            index=index,
+            columns=column_name,
+        )
+
     elif isinstance(model_component, pyo.Constraint):
         if dual_only:
             df = pd.DataFrame.from_dict(
-                {idx: {"Dual": model_component[idx].get_suffix_value("dual")} for idx in model_component},
+                {
+                    idx: {"Dual": model_component[idx].get_suffix_value("dual", default=np.nan)}
+                    for idx in model_component
+                },
                 orient="index",
             )
+            df.columns = column_name
         else:
+            # todo: why do these get special treatment?
             constraints_to_print_expr = [
                 "Rep_Period_Energy_Budget_Constraint",
             ]
@@ -117,17 +114,15 @@ def convert_pyomo_object_to_dataframe(
                         "Lower Bound": pyo.value(model_component[idx].lower),
                         "Body": pyo.value(model_component[idx].body),
                         "Upper Bound": pyo.value(model_component[idx].upper),
-                        "Dual": model_component[idx].get_suffix_value("dual"),
-                        "Expression": (
-                            model_component[idx].expr if model_component.name in constraints_to_print_expr else None
-                        ),
+                        "Dual": model_component[idx].get_suffix_value("dual", default=np.nan),
+                        "Expression": model_component[idx].expr
+                        if model_component.name in constraints_to_print_expr
+                        else None,
                     }
                     for idx in model_component
                 },
                 orient="index",
             )
-        # TODO (2022-02-22): The way the index is created and named for constraints vs. other model components seems
-        #  like could be made to be the same (related to previous TODO)
         index_names = get_index_labels(model_component)
         # If DataFrame is empty, need an extra step to be able to label the index headers
         if df.empty:
